+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+
+- Fix secuirty vulnerabilities from project dev dependencies
+
 ## [2.17.0] - 2019-10-28
 
 ### Added