+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Changed
-- Remove call to getSession in first component, use sessionPromise from render.
+- Remove call to `getSession` in first component, use sessionPromise from render.
 
 ### Fixed